Fix all compiler warnings
- Change next_cpu_id to atomic_int type for proper atomic_fetch_add usage
- Remove unused variable 'region' in pmm_add_region
- Mark unused function remove_free_block with __attribute__((unused))
- Cast pointer subtraction to uint64_t for signed/unsigned comparison

+++ kernel/multiboot2.c
@@ -156,7 +156,7 @@ void multiboot2_parse(uint32_t mbi_addr) {
         tag = (multiboot_tag_t *)((uint8_t *)tag + ((tag->size + 7) & ~7));
 
         /* Safety check - don't read beyond total_size */
-        if ((uint8_t *)tag - (uint8_t *)mbi >= total_size) {
+        if ((uint64_t)((uint8_t *)tag - (uint8_t *)mbi) >= total_size) {
             break;
         }
     }
